Clean up docs, add unit tests for SQLAlchemy
