+++ sql/game.py
@@ -9,8 +9,8 @@ class GameEntry(Base):
 
     game_date = Column(String, primary_key=True)
     game_time = Column(String, primary_key=True)
-    home_team = Column(String, primary_key=True)
-    away_team = Column(String, primary_key=True)
+    home_team = Column(String)
+    away_team = Column(String)
 
     ump_ks_per_game = Column(Float)
     ump_runs_per_game = Column(Float)
@@ -18,11 +18,8 @@ class GameEntry(Base):
     park_pitcher_score = Column(Float)
     wind_speed = Column(Integer)
 
-    """pitcher_postgame_entries = relationship('PostgamePitcherGameEntry', backref='game_entry')
-    hitter_postgame_entries = relationship('PostgameHitterGameEntry', backref='game_entry')
     pitcher_pregame_entries = relationship('PregamePitcherGameEntry', backref='game_entry')
-    pitcher_postgame_entries = relationship('PregameHitterGameEntry', backref='game_entry')
-"""
+    pitcher_postgame_entries = relationship('PostgamePitcherGameEntry', backref='game_entry')
 
     def __init__(self, game_date, game_time, home_team, away_team):
         self.game_date = game_date
