+++ sql/postgame_hitter.py
@@ -1,6 +1,6 @@
 
 from mlb_database import Base
-from sqlalchemy import Column, Integer, String, Float, ForeignKey
+from sqlalchemy import Column, Integer, String, Float, ForeignKey, ForeignKeyConstraint
 from sqlalchemy.orm import relationship
 
 
@@ -11,8 +11,11 @@ class PostgameHitterGameEntry(Base):
 
     rotowire_id = Column(String, ForeignKey('hitter_entries.rotowire_id'), primary_key=True)
     pitcher_id = Column(String, ForeignKey('pitcher_entries.rotowire_id'))
-    game_date = Column(String, ForeignKey('game_entries.game_date'), primary_key=True)
-    game_time = Column(String, ForeignKey('game_entries.game_time'), primary_key=True)
+    game_date = Column(String, primary_key=True)
+    game_time = Column(String, primary_key=True)
+    __table_args__ = (ForeignKeyConstraint([game_date, game_time],
+                                           ['game_entries.game_date', 'game_entries.game_time']), {})
+
     team = Column(String)
     opposing_team = Column(String)
     actual_draftkings_points = Column(Float)
