+++ sql/postgame_pitcher.py
@@ -1,6 +1,6 @@
 from sqlalchemy import Column, Integer, String, Float
 from mlb_database import Base
-from sqlalchemy import ForeignKey
+from sqlalchemy import ForeignKey, ForeignKeyConstraint
 
 
 class PostgamePitcherGameEntry(Base):
@@ -9,10 +9,12 @@ class PostgamePitcherGameEntry(Base):
     """
     __tablename__ = 'postgame_pitcher_entries'
 
-    rotowire_id = Column(String, ForeignKey('hitter_entries.rotowire_id'), primary_key=True)
-    pitcher_id = Column(String, ForeignKey('pitcher_entries.rotowire_id'))
-    game_date = Column(String, ForeignKey('game_entries.game_date'), primary_key=True)
-    game_time = Column(String, ForeignKey('game_entries.game_time'), primary_key=True)
+    rotowire_id = Column(String, ForeignKey('pitcher_entries.rotowire_id'), primary_key=True)
+    game_date = Column(String, primary_key=True)
+    game_time = Column(String, primary_key=True)
+    __table_args__ = (ForeignKeyConstraint([game_date, game_time],
+                                           ['game_entries.game_date', 'game_entries.game_time']), {})
+
     team = Column(String)
     opposing_team = Column(String)
     actual_draftkings_points = Column(Float)
