+++ sql/pregame_pitcher.py
@@ -12,8 +12,11 @@ class PregamePitcherGameEntry(Base):
     __tablename__ = 'pregame_pitcher_entries'
 
     rotowire_id = Column(String, ForeignKey('pitcher_entries.rotowire_id'), primary_key=True)
-    game_date = Column(String, ForeignKey('game_entries.game_date'), primary_key=True)
-    game_time = Column(String, ForeignKey('game_entries.game_time'), primary_key=True)
+    game_date = Column(String, primary_key=True)
+    game_time = Column(String, primary_key=True)
+    __table_args__ = (ForeignKeyConstraint([game_date, game_time],
+                                           ['game_entries.game_date', 'game_entries.game_time']), {})
+
     team = Column(String)
     opposing_team = Column(String)
     predicted_draftkings_points = Column(Float)
